Don't pass SSO session id in URL parameter.
There is no need to pass the SSO session_id in an URL parameter when
it is also sent to the users browser as a cookie.

In fact, it is preferable to NOT have it in the URL since these URLs
are bound to be put in an e-mail (perhaps to support) sooner or later.