add IdPContext
Put the things we need 'everywhere' in a separate context dataclass
and pass that down the stack rather than passing the top-level
IdPApplication.

This gives us the possiblity to do type checking without getting
circular imports where IdPApplication imports 'a' and 'a' imports
IdPApplication in order to verify it's type.

It should also make it easier to do the rewrite when moving from
CherryPy to Flask.